renamed bean package to domain
