+++ src/main/java/de/lh/tool/domain/dto/ExceptionDto.java
@@ -1,4 +1,4 @@
-package de.lh.tool.bean.dto;
+package de.lh.tool.domain.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
