+++ src/main/java/de/lh/tool/service/entity/impl/UserServiceImpl.java
@@ -2,7 +2,7 @@ package de.lh.tool.service.entity.impl;
 
 import org.springframework.stereotype.Service;
 
-import de.lh.tool.bean.model.User;
+import de.lh.tool.domain.model.User;
 import de.lh.tool.repository.UserRepository;
 import de.lh.tool.service.entity.interfaces.UserService;
