+++ src/main/java/de/lh/tool/service/entity/interfaces/UserService.java
@@ -1,6 +1,6 @@
 package de.lh.tool.service.entity.interfaces;
 
-import de.lh.tool.bean.model.User;
+import de.lh.tool.domain.model.User;
 
 public interface UserService extends BasicEntityService<User, Long> {
