+++ src/main/java/de/lh/tool/service/rest/exception/RestExceptionHandler.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
-import de.lh.tool.bean.dto.ExceptionDto;
+import de.lh.tool.domain.dto.ExceptionDto;
 
 @ControllerAdvice
 public class RestExceptionHandler extends ResponseEntityExceptionHandler {
